feat: 구간 등록 기능 구현
