+++ src/main/java/nextstep/subway/domain/Line.java
@@ -1,0 +1,66 @@
+package nextstep.subway.domain;
+
+import javax.persistence.*;
+import java.util.List;
+
+@Entity
+public class Line extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String name;
+
+    @Column(nullable = false)
+    private String color;
+
+    @Embedded
+    private Sections sections = new Sections();
+
+
+    protected Line() {
+
+    }
+
+    public Line(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public Line(String name, String color, Section section) {
+        this.name = name;
+        this.color = color;
+        addSection(section);
+    }
+
+    public void addSection(Section section) {
+        sections.addSection(section);
+        section.setLine(this);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public void changeName(String name) {
+        this.name = name;
+    }
+
+    public void changeColor(String color) {
+        this.color = color;
+    }
+
+    public List<Station> getStations() {
+        return sections.getStations();
+    }
+
+}
