--- src/main/java/nextstep/subway/domain/Line/Line.java
@@ -1,80 +1,0 @@
-package nextstep.subway.domain.Line;
-
-import nextstep.subway.domain.BaseEntity;
-import nextstep.subway.domain.station.Station;
-
-import javax.persistence.*;
-import java.util.Arrays;
-import java.util.List;
-
-@Entity
-public class Line extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(unique = true)
-    private String name;
-
-    @Column(nullable = false)
-    private String color;
-
-    @ManyToOne
-    @JoinColumn(name = "up_station_id")
-    private Station upStation;
-
-    @ManyToOne
-    @JoinColumn(name = "down_station_id")
-    private Station downStation;
-
-    @Column(nullable = false)
-    private int distance;
-
-    protected Line() {
-
-    }
-
-    public Line(String name, String color, Station upStation, Station downStation, int distance) {
-        this.name = name;
-        this.color = color;
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.distance = distance;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
-    }
-
-    public List<Station> getStations() {
-        return Arrays.asList(upStation, downStation);
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-
-    public void changeName(String name){
-        this.name = name;
-    }
-
-    public void changeColor(String color){
-        this.color = color;
-    }
-}
